Add files via upload
The file cv_dpdadaptLASSO_samelambda contains the main function. As arguments we need to introduce the data (matrices X and Y), an initializer mode (between RLARS or standar DPD-LASSO), a weight.mode (between "lasso", "adapt" and "scad" for the three differents weights functions used in the main paper) and a mode of choosing the lambda grid. "lambda0 uses an auxiliare function to determine the extremes of the path, and "given" uses the "lmin" and "lmax" limits provided by the user. Finally, "nlambda" determines the length of the lambda grid.
The tunning parameter (lambda) selection is performed using the HBIC criterion (see the main paper for more details).